requirements.txt
# Core
SQLAlchemy>=2.0
psycopg>=3.1

# Scheduler
APScheduler>=3.10

# HTTP client
httpx>=0.27

# Utils
python-dotenv
opencv-python>=4.9
numpy>=1.26
ultralytics>=8.1
